FogBugz #178128: Conditionally enable L2 EDAC.
Currently the L2 cache is the only ECC module implemented in
the ECC Manager. However, as more modules are added, the need
to turn each module on & off is required. The addition of
conditional compiles based on a new CONFIG (L2 ECC) is added
here.

V2:
Include dependency on CONFIG_CACHE_L2X0

Signed-off-by: Thor Thayer <tthayer@altera.com>